refactor: Mover el hook useSales a su propio archivo para mejorar la organización del código
